@@ -8,7 +8,7 @@
   "scripts": {
     "build": "rollup -c",
     "watch": "rollup -c -w",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "cypress run"
   },
   "repository": {
     "type": "git",
@@ -27,9 +27,11 @@
   "homepage": "https://github.com/koole/react-sanctum#readme",
   "devDependencies": {
     "@rollup/plugin-json": "^4.1.0",
-    "@types/react": "^16.14.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "@types/react": "^17.0.18",
+    "@types/react-dom": "^17.0.9",
+    "cypress": "^8.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "rollup": "^2.32.1",
     "rollup-plugin-typescript2": "^0.27.3",
     "typescript": "^3.9.7"
